Move StudentList component to shared folder for instructor view
