+++ app/assets/javascripts/components/students/shared/StudentList/StudentRow.jsx
